Create sd_rolling_files.ino
How the Code Works:Setup:Initializes the SD card.Sets up the maximum number of files (maxFiles) to store on the SD card.Loop:The loop function contains a dummy array dataArray that gets stored in the SD card using the storeData function. The delay is added for demonstration.storeData Function:This function first finds the latest file index using the findLatestFileIndex function.If the number of files exceeds the maximum allowed (maxFiles), the oldest file is removed, and the remaining files are renamed in a rolling fashion.A new file is then created with an incremented index, and the data is stored in this file.fileName Function:This function generates the file name in hexadecimal format (e.g., 0.txt, 1.txt, etc.).findLatestFileIndex Function:This function finds the index of the latest file stored on the SD card. It returns -1 if no files are found.Usage:Adjust maxFiles to the desired number of files to be stored on the SD card.The dataArray and dataSize in the loop function can be replaced with the actual data you wish to store.The code automatically handles rolling over and renaming files once the maximum number is exceeded.